Squashed 'vendor/github.com/htop-dev/htop/' changes from 2698483444..4abe9f4ce5
4abe9f4ce5 Update Settings.c
fdbb0629bc Minor bug fixes for Coverity issues
541c17c975 darwin/Platform.c: use mach/port.h
32cb30273b Disable basename matching for kernel threads
ea30b98b86 Read executable name before checking the command line
1f664ee214 Properly track if we actually hit any token while scanning cmdline
94c7822401 Explicitly check sscanf(3) and fscanf(3) return values
207db2e8f8 Linux: drop preconditions for free_and_xStrdup()
8e11f1f656 Linux: rename variable for clearer intent
85e4e02193 RichString: add access function annotations
db65f175ab Handle invalid multi-byte strings gracefully
e401c8fed9 Fix usage of RichString_appendnWide()
2f67a2b07d DragonFlyBSD: support Disk IO meter
1414cdcfc6 DragonFlyBSD: support Network IO meter
7639d62459 DragonFlyBSD: build fixes
91990b1a34 Clicking on column headers needs to also update the sort direction arrow
5d778eaacc Undo too-aggressive code collapsing from tree mode refactoring
49c13511c4 CRT: print message with pending changes only
8e40d68867 CRT: print a warning when exiting on signal
b6b9384678 Fix the display of number of running tasks
046b6ce466 fixes for illumos/Solaris platforms
e59176f23c Fix static linking with libsystemd
2a58f5dba8 Add -dev prefix again for the next (planned) release
68c970c7ef Release 3.3.0
2b6304cec5 Linux: use fseek(3)
f9fb93aaf8 Refactor the goThroughEntries.
615d489597 Fix build in NetBSD.
71021b2d8b Attempt to "Fix" OpenBSD Build
7bf8f02de2 Clarify isNonnegative() behavior in comments
c3b4a770c0 Updates for 2024. Happy New Year!
b149a506cf Shorten docker container IDs to 12 characters
955208569c Settings: preserve empty header
887c4eaff3 Fix execlp() argument without pointer cast
e5c09ffd96 Row_printTime() code shrink
e9cf230aa0 Row_printKBytes() improvements
c845b79ea7 Clarify magic number for data/cols fields
338baa4173 Fix naming convention inconsistency
4bcea76663 OpenFilesScreen: Make column sizing dynamic for file size, offset and inode
c55401b82b Add support for "truss" (FreeBSD equivalent of "strace")
7ee92be6fd Darwin: add NetworkIOMeter support
6490590dc8 Relocate include of config.h from header to source modeule
cf9abeb695 Document compilation issues and changed config.h inclusion rules
acb1ae4d29 Minor code simplification
2af66a4aba Convince IWYU to keep size_t/ssize_t via stddef.h
9ee8a7ca64 Force to keep explicit forward declarations
e56089e1fa Update includes based on IWYU
20bd78493b Update list of special header mappings for IWYU
d0a57b61c1 Force configuration with clang for IWYU runs
2aacbf84cb Use consistent style for process field output/compare functions
6aa9ef2726 Fix code style
a98fc47ffe Check for large file support and some types we use
5c7cb91d18 Do not touch CFLAGS in configure.ac
179aeb0108 Add includes for config.h as per the discussion in PR #1337
8980d7a896 HeaderLayout: add "3 columns - 40/30/30", "... 30/40/30" & "... 30/30/40"
f1f69d6d1c Meter: use correct unicode characters for digit '9'
170354c638 Bump github/codeql-action from 2 to 3
f55b7a6e1d Merge branch 'patch-2' (Update manual to mention KiB) of sparrowt/htop
23b0bdf464 Review comments for memory size manual clarifications
cabb95ea92 Note in manual re default memory units of KiB
92a68c097f Consolidate duplicate string constants in CGroup filters
e64adcb50b Add PCP support
d5b1556497 Add column for process container name
1069e069ab Add logic to filter the container name (+type) from the CGroup name
3d0f35c0b7 Merge branch 'main' of https://github.com/BowDown097/htop into BowDown097-main
8915caedc5 Fix a little header comment typo in the CGroupUtils code
29137ccc7c PCP platform implementation of the compressed cgroup column
0600ef50f5 Update pcp/ProcessField.h
d20fcab197 Update linux/LinuxProcessTable.c
ed1c7450e2 Use printKBytes
986dbfc19c Linux + PCP: Add private memory size column
1d3a823d5d Change NetworkIOMeter value unit from KiB/s to bytes/second
fda3d10261 Cache NetworkIOMeter values in 'double' type
d474689689 NetworkIOMeter: Cache the formatted "KiB/s" number strings
5affa05bd6 DiskIOMeter: Cache the formatted "KiB/s" number strings
5d2569270c Use xSnprintf in Meter_humanUnit()
5751bafb8d Rewrite Meter_humanUnit() to accept floating point value
e34a9fcc8c Always update 'values' member of NetworkIOMeter
f7f6aca39d Always update 'values' member of DiskIOMeter
9352e26218 NetworkIOMeter: correct format string and code comments
7d7e8df83a Use xSnprintf in DiskIOMeter_updateValues()
89b90c4e41 Cap DiskIOMeter "utilisation" percentage at 100%

git-subtree-dir: vendor/github.com/htop-dev/htop
git-subtree-split: 4abe9f4ce5102d029f65d26e73b03b00005a6096

+++ pcp/PCPDynamicMeter.c
@@ -351,27 +351,27 @@ void PCPDynamicMeter_updateValues(PCPDynamicMeter* this, Meter* meter) {
             break;
          case PM_TYPE_32:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.l, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.l, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%d", atom.l);
             break;
          case PM_TYPE_U32:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.ul, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.ul, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%u", atom.ul);
             break;
          case PM_TYPE_64:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.ll, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.ll, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%lld", (long long) atom.ll);
             break;
          case PM_TYPE_U64:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.ull, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.ull, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%llu", (unsigned long long) atom.ull);
             break;
          case PM_TYPE_FLOAT:
             bytes += conv.dimSpace ?
-               Meter_humanUnit(buffer + bytes, atom.f, size - bytes) :
+               Meter_humanUnit(buffer + bytes, (double) atom.f, size - bytes) :
                xSnprintf(buffer + bytes, size - bytes, "%.2f", (double) atom.f);
             break;
          case PM_TYPE_DOUBLE:
@@ -382,9 +382,11 @@ void PCPDynamicMeter_updateValues(PCPDynamicMeter* this, Meter* meter) {
          default:
             break;
       }
+
       if (saved != bytes && metric->suffix)
          bytes += xSnprintf(buffer + bytes, size - bytes, "%s", metric->suffix);
    }
+
    if (!bytes)
       xSnprintf(buffer, size, "no data");
 }
@@ -427,27 +429,27 @@ void PCPDynamicMeter_display(PCPDynamicMeter* this, ATTR_UNUSED const Meter* met
             break;
          case PM_TYPE_32:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.l, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.l, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%d", atom.l);
             break;
          case PM_TYPE_U32:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.ul, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.ul, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%u", atom.ul);
             break;
          case PM_TYPE_64:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.ll, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.ll, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%lld", (long long) atom.ll);
             break;
          case PM_TYPE_U64:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.ull, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.ull, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%llu", (unsigned long long) atom.ull);
             break;
          case PM_TYPE_FLOAT:
             len = conv.dimSpace ?
-               Meter_humanUnit(buffer, atom.f, sizeof(buffer)) :
+               Meter_humanUnit(buffer, (double) atom.f, sizeof(buffer)) :
                xSnprintf(buffer, sizeof(buffer), "%.2f", (double) atom.f);
             break;
          case PM_TYPE_DOUBLE:
@@ -458,12 +460,14 @@ void PCPDynamicMeter_display(PCPDynamicMeter* this, ATTR_UNUSED const Meter* met
          default:
             break;
       }
+
       if (len) {
          RichString_appendnAscii(out, CRT_colors[metric->color], buffer, len);
          if (metric->suffix)
             RichString_appendAscii(out, CRT_colors[METER_TEXT], metric->suffix);
       }
    }
+
    if (nodata)
       RichString_writeAscii(out, CRT_colors[METER_VALUE_ERROR], "no data");
 }
